backend/requirements.txt
flask
influxdb-client[ciso]
python-dotenv